requirement.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-migrate
flask-bcrypt
flask-jwt-extended